extract drawer context to a different file
